Distribute all static content
